Extract ParameterVariable from ThisVariable
